moving NftMarkerCreator libs to libs folder and changes of the paths
