+++ docs/index.html
@@ -50,9 +50,9 @@
     <!-- <script src="PNGReader.js"></script> -->
     <script src="exif.js"></script>
     <script type='text/javascript'>
-          var artoolkit_wasm_url = '../NftMarkerCreator_wasm.wasm';
+          var artoolkit_wasm_url = '../libs/NftMarkerCreator_wasm.wasm';
     </script>
-    <script async type="text/javascript" src="../NftMarkerCreator_wasm.js"></script>
+    <script async type="text/javascript" src="../libs/NftMarkerCreator_wasm.js"></script>
     <script src="index.js"></script>
 </body>
 </html>
