+++ libs/NftMarkerCreator_wasm.js
