+++ libs/NftMarkerCreator_wasm.wasm
